feat(selection): hold the best clips longer, or finish short, rather than pad
When backfill ran out of anything admissible it relaxed its constraints until
something fit, and a clip nobody would have chosen ended up in the cut to buy
a few seconds. Two better answers, in order.

Hold the strongest clips a little longer — up to two seconds each, never past
the end of their source. Two more seconds of the best clip answers a
four-second gap better than a weak clip that earned no place in the memory.

Then, if it is still short, let it be short. A recap that ends early beats one
padded with material the pool could not justify, and the log says so rather
than leaving a quietly worse video to explain itself.

The backfill machinery moves to analysis/clip_backfill.py — clip_refiner.py
had passed the 1000-line hard limit again, and "how a cut is filled to its
runtime" is a different job from "what belongs in it". The order of
concessions is the whole design of that module, so it now says so at the top.
