+++ complexipy-snapshot.json
@@ -6,101 +6,101 @@
       {
         "name": "ClipRefiner::_fill_empty_weeks",
         "complexity": 20,
-        "line_start": 527,
-        "line_end": 570,
+        "line_start": 234,
+        "line_end": 277,
         "line_complexities": [
           {
-            "line": 537,
+            "line": 244,
             "complexity": 0
           },
           {
-            "line": 538,
+            "line": 245,
             "complexity": 1
           },
           {
-            "line": 539,
+            "line": 246,
             "complexity": 0
           },
           {
-            "line": 542,
+            "line": 249,
             "complexity": 0
           },
           {
-            "line": 543,
+            "line": 250,
             "complexity": 1
           },
           {
-            "line": 544,
+            "line": 251,
             "complexity": 0
           },
           {
-            "line": 547,
+            "line": 254,
             "complexity": 0
           },
           {
-            "line": 548,
+            "line": 255,
             "complexity": 0
           },
           {
-            "line": 550,
+            "line": 257,
             "complexity": 0
           },
           {
-            "line": 551,
+            "line": 258,
             "complexity": 0
           },
           {
-            "line": 552,
+            "line": 259,
             "complexity": 1
           },
           {
-            "line": 553,
+            "line": 260,
             "complexity": 0
           },
           {
-            "line": 554,
+            "line": 261,
             "complexity": 2
           },
           {
-            "line": 556,
+            "line": 263,
             "complexity": 0
           },
           {
-            "line": 558,
+            "line": 265,
             "complexity": 0
           },
           {
-            "line": 559,
+            "line": 266,
             "complexity": 1
           },
           {
-            "line": 560,
+            "line": 267,
             "complexity": 2
           },
           {
-            "line": 561,
+            "line": 268,
             "complexity": 0
           },
           {
-            "line": 562,
+            "line": 269,
             "complexity": 3
           },
           {
-            "line": 564,
+            "line": 271,
             "complexity": 4
           },
           {
-            "line": 565,
+            "line": 272,
             "complexity": 5
           },
           {
-            "line": 570,
+            "line": 277,
             "complexity": 0
           }
         ]
       }
     ],
-    "complexity": 121
+    "complexity": 97
   },
   {
     "path": "src/immich_memories/analysis/unified_budget.py",
